# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bible API
  description: |
    Complete Bible API with Role-Based Access Control (RBAC), multi-version support, RSS feeds, user management, file locations, and backup system.
    
    ## Role-Based Access Control
    - **Admin**: Full system access with all permissions
    - **User**: Standard user with Bible reading and format access
    - **Viewer**: Read-only access to Bible content
    - **Editor**: Can edit content but not manage users
    
    ## Authentication
    Use Bearer token authentication. Login to get your JWT token.
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 attempts per 15 minutes
    
    ## File Locations
    Users can configure local drive paths where TXT and XML files will be saved automatically when accessing verses.
  version: 3.0.0
  contact:
    name: API Support
    email: support@bibleapi.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.bibleapi.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and get JWT token with session
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
              properties:
                user_id:
                  type: string
                  minLength: 3
                  maxLength: 50
                  example: admin
                password:
                  type: string
                  format: password
                  minLength: 4
                  example: admin123
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT token for authentication
                      session_id:
                        type: string
                        description: Unique session identifier
                      session_expiry:
                        type: string
                        format: date-time
                      user:
                        type: object
                        properties:
                          user_id:
                            type: string
                          email:
                            type: string
                          full_name:
                            type: string
                          role:
                            type: string
                      versions:
                        type: array
                        items:
                          type: string
                      formats:
                        type: array
                        items:
                          type: string
                      permissions:
                        type: array
                        items:
                          type: string
                      expires_in:
                        type: string
        400:
          description: Missing credentials
        401:
          description: Invalid credentials or account inactive
        429:
          description: Too many login attempts
  /auth/logout:
    post:
      summary: User logout
      description: Logout user and clear session history
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        401:
          description: Authentication required
  /auth/check-session:
    get:
      summary: Check session validity
      description: Verify if current session is still valid
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Session is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        401:
          description: Session expired or invalid

  # ==================== USER INFORMATION ====================
...
  /api/{version}/books:
    get:
      summary: Get books list
      description: Returns list of all books in the Bible with testament classification
      tags: [Bible]
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of books
        401:
          description: Authentication required
  /api/books/info/{bookCode}:
    get:
      summary: Get book information
      description: Returns detailed information about a specific book
      tags: [Bible]
      security:
        - bearerAuth: []
      parameters:
        - name: bookCode
          in: path
          required: true
          schema:
            type: string
          example: PSA
      responses:
        200:
          description: Book information
        404:
          description: Book not found
...
  /admin/users:
    get:
      summary: Get all users
      description: Get list of all users with their roles (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
        403:
          description: Admin access required
  /admin/users/{userId}:
    get:
      summary: Get user by ID
      description: Get detailed information about a specific user (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User details
        403:
          description: Admin access required
        404:
          description: User not found

    patch:
      summary: Update user
      description: Update user information (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                full_name:
                  type: string
                is_active:
                  type: boolean
      responses:
        200:
          description: User updated
        403:
          description: Admin access required
        404:
          description: User not found
  /admin/users/{userId}/toggle-status:
    patch:
      summary: Activate/Deactivate user
      description: Toggle user active status (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User status updated
        403:
          description: Admin access required
        404:
          description: User not found
  /admin/stats:
    get:
      summary: Get system statistics
      description: Get overall system statistics (Admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        200:
          description: System statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      users:
                        type: integer
                      api_calls:
                        type: integer
                      active_users:
                        type: integer
        403:
          description: Admin access required
...